Refactor runner
