+++ main/main.go
@@ -8,12 +8,15 @@ import (
 	"errors"
 	"fmt"
 	"os"
+	"os/signal"
+	"syscall"
 
 	"github.com/spf13/pflag"
 	"golang.org/x/term"
 
 	"github.com/ava-labs/avalanchego/config"
-	"github.com/ava-labs/avalanchego/node/bootstrap"
+	"github.com/ava-labs/avalanchego/node"
+	"github.com/ava-labs/avalanchego/utils"
 	"github.com/ava-labs/avalanchego/version"
 )
 
@@ -60,11 +63,60 @@ func main() {
 		fmt.Println(Header)
 	}
 
-	runner, err := bootstrap.New(nodeConfig)
+	runner, err := node.NewRunner(nodeConfig)
 	if err != nil {
 		fmt.Printf("couldn't start node: %s\n", err)
 		os.Exit(1)
 	}
 
 	os.Exit(Run(runner))
+}
+
+const Header = `     _____               .__                       .__
+    /  _  \___  _______  |  | _____    ____   ____ |  |__   ____    ,_ o
+   /  /_\  \  \/ /\__  \ |  | \__  \  /    \_/ ___\|  |  \_/ __ \   / //\,
+  /    |    \   /  / __ \|  |__/ __ \|   |  \  \___|   Y  \  ___/    \>> |
+  \____|__  /\_/  (____  /____(____  /___|  /\___  >___|  /\___  >    \\
+          \/           \/          \/     \/     \/     \/     \/`
+
+func Run(app *node.Runner) int {
+	// start running the application
+	app.Start()
+
+	// register terminationSignals to kill the application
+	terminationSignals := make(chan os.Signal, 1)
+	signal.Notify(terminationSignals, syscall.SIGINT, syscall.SIGTERM)
+
+	stackTraceSignal := make(chan os.Signal, 1)
+	signal.Notify(stackTraceSignal, syscall.SIGABRT)
+
+	// start up a new go routine to handle attempts to kill the application
+	go func() {
+		for range terminationSignals {
+			app.Stop()
+			return
+		}
+	}()
+
+	// start a goroutine to listen on SIGABRT signals,
+	// to print the stack trace to standard error.
+	go func() {
+		for range stackTraceSignal {
+			fmt.Fprint(os.Stderr, utils.GetStacktrace(true))
+		}
+	}()
+
+	// wait for the app to exit and get the exit code response
+	exitCode := app.ExitCode()
+
+	// shut down the termination signal go routine
+	signal.Stop(terminationSignals)
+	close(terminationSignals)
+
+	// shut down the stack trace go routine
+	signal.Stop(stackTraceSignal)
+	close(stackTraceSignal)
+
+	// return the exit code that the application reported
+	return exitCode
 }
